email-telefonszám módosítás (telszám még nemjó)

+++ project/Frontend/js/controllers/settingsCtrl.js
@@ -1,7 +1,9 @@
 app.controller('settingsCtrl', function ($scope, $rootScope, factory, $location) {
     $scope.userPerm = ['inactive', 'student', 'teacher', 'admin'];
     $scope.password = {};
-    $scope.phonenumber = [];
+    $scope.phoneNum = {};
+    $scope.email = {};
+    
 
     $scope.passMod = function () {
         if ($scope.password.password == null || $scope.uNewPasswordAgain == null) {
@@ -15,7 +17,29 @@ app.controller('settingsCtrl', function ($scope, $rootScope, factory, $location)
                     let data = { password: CryptoJS.SHA1($scope.password.password).toString() };
                     console.log(data);
                     factory.update($scope.userPerm[$rootScope.loggedPermission], res[0].ID, data).then(function (res) {
-                        alert('Sikeres jelszóváltozatás!');
+                        alert('Sikeres jelszó változatás!');
+                        sessionStorage.clear();
+                        $rootScope.loggedIn = false;
+                        $location.path('#!/');
+                    });
+                });
+            }
+        }
+    };
+
+    $scope.phoneMod = function () {
+        if ($scope.phoneNum.phoneNum == null || $scope.uNewPhoneAgain == null) {
+            alert('Nem adott meg minden adatot!');
+        } else {
+            if ($scope.phoneNum.phoneNum != $scope.uNewPhoneAgain) {
+                alert('A két email nem egyezik meg!');
+            } else {
+                factory.select($scope.userPerm[$rootScope.loggedPermission], 'email', $rootScope.loggedUserMail).then(function (res) {
+                    console.log(res);
+                    let data = { phoneNum: $scope.phoneNum.phoneNum};
+                    console.log(data);
+                    factory.update($scope.userPerm[$rootScope.loggedPermission], res[0].ID, data).then(function (res) {
+                        alert('Sikeres telefonszám változtatás!');
                         sessionStorage.clear();
                         $rootScope.loggedIn = false;
                         $location.path('#!/');
